requirements.txt
# Core Framework
streamlit>=1.31.0
python-dotenv>=1.0.0

# Document Processing
pdfplumber>=0.10.3
PyPDF2>=3.0.1
python-docx>=1.1.0

# NLP Processing
spacy>=3.7.2
nltk>=3.8.1
langdetect>=1.0.9

# LLM Integration
openai>=1.12.0
anthropic>=0.18.0

# PDF Report Generation
fpdf2>=2.7.6

# Data Processing
pandas>=2.1.4
numpy>=1.26.3

# Testing
pytest>=7.4.4
